sui_indexer_alt_metrics
Struct MetricsService

pub struct MetricsService { /* private fields */ }

A service that exposes prometheus metrics over HTTP on a “/metrics” route on the provided listen address.

Implementations

impl MetricsService

pub fn new( args: MetricsArgs, registry: Registry, cancel: CancellationToken, ) -> Self

Create a new instance of the service, listening on the address provided in args, serving metrics from the registry. The service will shut down if the provided cancel token is cancelled.

The service will not be run until Self::run is called.

pub fn registry(&self) -> &Registry

Add metrics to this registry to serve them from this service.

pub async fn run(self) -> Result<JoinHandle<()>>

Start the service. The service will run until the cancellation token is triggered.